+++ Socolin.RabbitMQ.Client/Options/Consumer/DeserializationPipeOptions.cs
@@ -4,11 +4,11 @@ namespace Socolin.RabbitMQ.Client.Options.Consumer
 {
 	public class DeserializationPipeOptions<T> where T : class
 	{
-		public readonly SerializerDelegate<T> DefaultDeserializer;
+		public readonly SerializerDelegate<T>? DefaultDeserializer;
 		public readonly IReadOnlyDictionary<string, SerializerDelegate<T>> Deserializers;
 
 		public DeserializationPipeOptions(
-			SerializerDelegate<T> defaultDeserializer,
+			SerializerDelegate<T>? defaultDeserializer,
 			IReadOnlyDictionary<string, SerializerDelegate<T>> deserializers
 		)
 		{